requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
dateutils==0.6.12
et-xmlfile==1.1.0
idna==3.3
nsetools==1.0.11
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
pynse==0.1.5
pypiwin32==223
pyreadline==2.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
urllib3==1.26.9
XlsxWriter==3.0.3
xlwings==0.27.6
git+https://github.com/StreamAlpha/pynse.git@master